Deprecate level keyword for dataframe and series aggregations (#40869)
