+++ pandas/tests/series/methods/test_count.py
@@ -19,13 +19,15 @@ class TestSeriesCount:
 
         ser = Series(np.random.randn(len(index)), index=index)
 
-        result = ser.count(level=0)
+        with tm.assert_produces_warning(FutureWarning):
+            result = ser.count(level=0)
         expected = ser.groupby(level=0).count()
         tm.assert_series_equal(
             result.astype("f8"), expected.reindex(result.index).fillna(0)
         )
 
-        result = ser.count(level=1)
+        with tm.assert_produces_warning(FutureWarning):
+            result = ser.count(level=1)
         expected = ser.groupby(level=1).count()
         tm.assert_series_equal(
             result.astype("f8"), expected.reindex(result.index).fillna(0)
@@ -37,24 +39,30 @@ class TestSeriesCount:
         series = ser.copy()
         series.index.names = ["a", "b"]
 
-        result = series.count(level="b")
-        expect = ser.count(level=1).rename_axis("b")
+        with tm.assert_produces_warning(FutureWarning):
+            result = series.count(level="b")
+        with tm.assert_produces_warning(FutureWarning):
+            expect = ser.count(level=1).rename_axis("b")
         tm.assert_series_equal(result, expect)
 
-        result = series.count(level="a")
-        expect = ser.count(level=0).rename_axis("a")
+        with tm.assert_produces_warning(FutureWarning):
+            result = series.count(level="a")
+        with tm.assert_produces_warning(FutureWarning):
+            expect = ser.count(level=0).rename_axis("a")
         tm.assert_series_equal(result, expect)
 
         msg = "Level x not found"
         with pytest.raises(KeyError, match=msg):
-            series.count("x")
+            with tm.assert_produces_warning(FutureWarning):
+                series.count("x")
 
     def test_count_level_without_multiindex(self):
         ser = Series(range(3))
 
         msg = "Series.count level is only valid with a MultiIndex"
         with pytest.raises(ValueError, match=msg):
-            ser.count(level=1)
+            with tm.assert_produces_warning(FutureWarning):
+                ser.count(level=1)
 
     def test_count(self, datetime_series):
         assert datetime_series.count() == len(datetime_series)
@@ -66,12 +74,14 @@ class TestSeriesCount:
         mi = MultiIndex.from_arrays([list("aabbcc"), [1, 2, 2, np.nan, 1, 2]])
         ts = Series(np.arange(len(mi)), index=mi)
 
-        left = ts.count(level=1)
+        with tm.assert_produces_warning(FutureWarning):
+            left = ts.count(level=1)
         right = Series([2, 3, 1], index=[1, 2, np.nan])
         tm.assert_series_equal(left, right)
 
         ts.iloc[[0, 3, 5]] = np.nan
-        tm.assert_series_equal(ts.count(level=1), right - 1)
+        with tm.assert_produces_warning(FutureWarning):
+            tm.assert_series_equal(ts.count(level=1), right - 1)
 
         # GH#29478
         with pd.option_context("use_inf_as_na", True):
