+++ pandas/tests/series/test_reductions.py
@@ -63,7 +63,8 @@ def test_prod_numpy16_bug():
 def test_sum_with_level():
     obj = Series([10.0], index=MultiIndex.from_tuples([(2, 3)]))
 
-    result = obj.sum(level=0)
+    with tm.assert_produces_warning(FutureWarning):
+        result = obj.sum(level=0)
     expected = Series([10.0], index=[2])
     tm.assert_series_equal(result, expected)
